@@ -20,19 +20,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.16",
+    "@angular/animations": "8.2.14",
     "@angular/cdk": "^6.4.7",
-    "@angular/common": "7.2.16",
-    "@angular/compiler": "7.2.16",
-    "@angular/core": "7.2.16",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
     "@angular/flex-layout": "~7.0.0-beta.24",
-    "@angular/forms": "7.2.16",
-    "@angular/http": "7.2.16",
+    "@angular/forms": "8.2.14",
     "@angular/material": "^6.4.7",
     "@angular/material-moment-adapter": "^6.0.0",
-    "@angular/platform-browser": "7.2.16",
-    "@angular/platform-browser-dynamic": "7.2.16",
-    "@angular/router": "7.2.16",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "@ngx-translate/core": "10.0.2",
     "@ngx-translate/http-loader": "^3.0.1",
     "angular2-notifications": "1.0.4",
@@ -47,16 +46,16 @@
     "raven-js": "3.21.0",
     "rxjs": "6.6.3",
     "typeface-montserrat": "^0.0.45",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.4"
+    "web-animations-js": "^2.3.2",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.10",
-    "@angular/compiler-cli": "7.2.16",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "8.2.14",
     "@types/jasmine": "2.8.3",
     "@types/node": "~9.3.0",
-    "codelyzer": "^4.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "^2.8.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "^2.0.0",
@@ -72,6 +71,6 @@
     "ts-node": "^4.1.0",
     "tslint": "^5.9.1",
     "tslib": "^1.9.0",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   }
 }
