@@ -1,6 +1,6 @@
-import {Component, ContentChild, HostBinding, Input, ViewEncapsulation} from '@angular/core';
-import {OCIconDataChildComponent} from '../oc-icon-data-child/oc-icon-data-child.component';
-import {isTruthy} from '../../../../core/util/helper';
+import { Component, ContentChild, HostBinding, Input, ViewEncapsulation } from '@angular/core';
+import { OCIconDataChildComponent } from '../oc-icon-data-child/oc-icon-data-child.component';
+import { isTruthy } from '../../../../core/util/helper';
 
 @Component({
     selector: 'oc-icon-data',
@@ -13,7 +13,7 @@ export class OCIconDataComponent {
     @Input() public titleKey: string;
     @HostBinding('class.show-divider')
     @Input() public showDivider: boolean = true;
-    @ContentChild(OCIconDataChildComponent) child: OCIconDataChildComponent;
+    @ContentChild(OCIconDataChildComponent, { static: false }) child: OCIconDataChildComponent;
 
     public hasChild() {
         return isTruthy(this.child);
